changes template handling and data model
up until now, every few did call the ttemplate rendering on itself, this
clearly is a bad idea because every time it is called, a new context
for jinja has to be created.
The template handling has been moved to wurstgulasch.py, which
simplifies the handling a bit.
For every view there now has to be a template with by the name of
    < view name >.htmljinja
If a view has nothing to return it needs to return an empty dictionary.

Also, the data model has been changed, friends are now user-specific

oke das wa ales